+++ packages/components/src/components/input-date-picker/input-date-picker.tsx
@@ -1,6 +1,6 @@
 // @ts-strict-ignore
 import { isServer, PropertyValues } from "lit";
-import { createRef, Ref } from "lit-html/directives/ref.js";
+import { createRef, Ref } from "lit/directives/ref.js";
 import {
   createEvent,
   h,
@@ -307,6 +307,15 @@ export class InputDatePicker
 
   /** Specifies the status of the input field, which determines message and icons. */
   @property({ reflect: true }) status: Status = "idle";
+
+  /**
+   * When true, disables top layer placement when the component is open.
+   *
+   * Only set this if you need complex z-index control or if top layer placement causes conflicts with third-party components.
+   *
+   * @mdn [Top Layer](https://developer.mozilla.org/en-US/docs/Glossary/Top_layer)
+   */
+  @property({ reflect: true }) topLayerDisabled = false;
 
   /** Specifies the validation icon to display under the component. */
   @property({ reflect: true, converter: stringOrBoolean, type: String }) validationIcon: