+++ packages/components-react/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.0-next.71](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.70...@esri/calcite-components-react@5.0.0-next.71) (2026-01-16)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.70](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.69...@esri/calcite-components-react@5.0.0-next.70) (2026-01-16)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.69](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.68...@esri/calcite-components-react@5.0.0-next.69) (2026-01-16)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.68](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.67...@esri/calcite-components-react@5.0.0-next.68) (2026-01-16)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.67](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.66...@esri/calcite-components-react@5.0.0-next.67) (2026-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.66](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.65...@esri/calcite-components-react@5.0.0-next.66) (2026-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.65](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.64...@esri/calcite-components-react@5.0.0-next.65) (2026-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.64](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.63...@esri/calcite-components-react@5.0.0-next.64) (2026-01-15)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.63](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.62...@esri/calcite-components-react@5.0.0-next.63) (2026-01-14)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.62](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.61...@esri/calcite-components-react@5.0.0-next.62) (2026-01-14)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.61](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.60...@esri/calcite-components-react@5.0.0-next.61) (2026-01-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.60](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.59...@esri/calcite-components-react@5.0.0-next.60) (2026-01-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.59](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.58...@esri/calcite-components-react@5.0.0-next.59) (2026-01-13)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.58](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.57...@esri/calcite-components-react@5.0.0-next.58) (2026-01-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.57](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.56...@esri/calcite-components-react@5.0.0-next.57) (2026-01-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
+## [5.0.0-next.56](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.55...@esri/calcite-components-react@5.0.0-next.56) (2026-01-12)
+
+**Note:** Version bump only for package @esri/calcite-components-react
+
 ## [5.0.0-next.55](https://github.com/Esri/calcite-design-system/compare/@esri/calcite-components-react@5.0.0-next.54...@esri/calcite-components-react@5.0.0-next.55) (2026-01-09)
 
 **Note:** Version bump only for package @esri/calcite-components-react